backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.5.2
Flask-Cors==4.0.0
Flask-Migrate==4.0.5
python-dotenv==1.0.0
Werkzeug==2.3.7
psycopg2-binary==2.9.9
tabulate==0.9.0
qrcode==7.4.2
Pillow==10.2.0
razorpay==1.4.1
stripe==7.12.0
